LayoutTests/streams/readablestream-start.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script>

test(function()
{
    var isStartCalled = false;
    var source =
    { start:
        function(controller) {
            assert_equals(this, source);

            assert_array_equals(Object.getOwnPropertyNames(controller), ['enqueue', 'close', 'error']);

            enqueue = controller.enqueue;
            close = controller.close;
            error = controller.error;

            assert_equals(typeof enqueue, 'function');
            assert_equals(typeof close, 'function');
            assert_equals(typeof error, 'function');

            assert_array_equals(Object.getOwnPropertyNames(enqueue), ['name', 'length']);
            assert_array_equals(Object.getOwnPropertyNames(close), ['name', 'length']);
            assert_array_equals(Object.getOwnPropertyNames(error), ['name', 'length']);

            assert_equals(enqueue.name, '');
            assert_equals(close.name, '');
            assert_equals(error.name, '');

            assert_equals(enqueue.length, 1);
            assert_equals(close.length, 0);
            assert_equals(error.length, 1);

            isStartCalled = true;
        }
    };
    var rs = new ReadableStream(source);
    assert_true(isStartCalled);
}, 'ReadableStream start should be called with the proper parameters');

test(function()
{
    var isStartCalled = false;
    var source =
    { start:
        function(controller) {
            assert_array_equals(Object.getOwnPropertyNames(controller), ['enqueue', 'close', 'error']);
            controller.test = "";
            assert_array_equals(Object.getOwnPropertyNames(controller), ['enqueue', 'close', 'error', 'test']);

            isStartCalled = true;
        }
    };
    var rs = new ReadableStream(source);
    assert_true(isStartCalled);
}, 'ReadableStream start controller parameter should be updatable');

test(function()
{
    var isStartCalled = false;
    SimpleStreamSource = function() { };
    SimpleStreamSource.prototype.start = function() { isStartCalled = true; };
    SimpleStreamSource.prototype.constructor = SimpleStreamSource;

    var rs = new ReadableStream(new SimpleStreamSource());
    assert_true(isStartCalled);
}, 'ReadableStream should be able to call start method within prototype chain of its source');

test(function()
{
    mashedError = new Error('potato');
    assert_throws(mashedError, function() {
        var rs = new ReadableStream(
        { start:
            function() {
                throw mashedError;
            }
        });
    });
}, 'ReadableStream start should be able to throw');

test(function() {
    assert_throws(new TypeError(), function() {
        new ReadableStream({ start: 'potato'});
    });
}, 'ReadableStream constructor should get a function as start argument');

</script>
